move AtomicWriteFile and CurrentHomeDir() to osutil
