+++ coreconfig/config.go
@@ -32,7 +32,6 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/ubuntu-core/snappy/helpers"
 	"github.com/ubuntu-core/snappy/osutil"
 	"github.com/ubuntu-core/snappy/systemd"
 
@@ -323,7 +322,7 @@ var setTimezone = func(timezone string) error {
 		return err
 	}
 
-	return helpers.AtomicWriteFile(tzFile(), []byte(timezone), 0644, helpers.AtomicWriteFollow)
+	return osutil.AtomicWriteFile(tzFile(), []byte(timezone), 0644, osutil.AtomicWriteFollow)
 }
 
 func getPassthrough(rootDir string) (pc []passthroughConfig, err error) {
@@ -352,7 +351,7 @@ func setPassthrough(rootDir string, pc []passthroughConfig) error {
 			os.Remove(path)
 			continue
 		}
-		if err := helpers.AtomicWriteFile(path, []byte(c.Content), 0644, helpers.AtomicWriteFollow); err != nil {
+		if err := osutil.AtomicWriteFile(path, []byte(c.Content), 0644, osutil.AtomicWriteFollow); err != nil {
 			return err
 		}
 	}
@@ -388,7 +387,7 @@ var getModprobe = func() (string, error) {
 
 // setModprobe sets the specified modprobe config
 var setModprobe = func(modprobe string) error {
-	return helpers.AtomicWriteFile(modprobePath, []byte(modprobe), 0644, helpers.AtomicWriteFollow)
+	return osutil.AtomicWriteFile(modprobePath, []byte(modprobe), 0644, osutil.AtomicWriteFollow)
 }
 
 func getModules() ([]string, error) {
@@ -475,7 +474,7 @@ func setModules(modules []string) error {
 		buf.WriteByte('\n')
 	}
 
-	return helpers.AtomicWriteFile(modulesPath, buf.Bytes(), 0644, helpers.AtomicWriteFollow)
+	return osutil.AtomicWriteFile(modulesPath, buf.Bytes(), 0644, osutil.AtomicWriteFollow)
 }
 
 // getWatchdog returns the current watchdog config
@@ -501,11 +500,11 @@ var getWatchdog = func() (*watchdogConfig, error) {
 
 // setWatchdog sets the specified watchdog config
 var setWatchdog = func(wf *watchdogConfig) error {
-	if err := helpers.AtomicWriteFile(watchdogStartupPath, []byte(wf.Startup), 0644, helpers.AtomicWriteFollow); err != nil {
+	if err := osutil.AtomicWriteFile(watchdogStartupPath, []byte(wf.Startup), 0644, osutil.AtomicWriteFollow); err != nil {
 		return err
 	}
 
-	return helpers.AtomicWriteFile(watchdogConfigPath, []byte(wf.Config), 0644, helpers.AtomicWriteFollow)
+	return osutil.AtomicWriteFile(watchdogConfigPath, []byte(wf.Config), 0644, osutil.AtomicWriteFollow)
 }
 
 // for testing purposes
@@ -580,5 +579,5 @@ var setHostname = func(hostname string) error {
 		return err
 	}
 
-	return helpers.AtomicWriteFile(hostnamePath, hostnameB, 0644, helpers.AtomicWriteFollow)
+	return osutil.AtomicWriteFile(hostnamePath, hostnameB, 0644, osutil.AtomicWriteFollow)
 }
