+++ helpers/helpers_test.go
@@ -68,172 +68,6 @@ func (ts *HTestSuite) TestMakeRandomString(c *C) {
 	c.Assert(s2, Equals, "4PQyl")
 }
 
-func (ts *HTestSuite) TestAtomicWriteFile(c *C) {
-	tmpdir := c.MkDir()
-
-	p := filepath.Join(tmpdir, "foo")
-	err := AtomicWriteFile(p, []byte("canary"), 0644, 0)
-	c.Assert(err, IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "canary")
-
-	// no files left behind!
-	d, err := ioutil.ReadDir(tmpdir)
-	c.Assert(err, IsNil)
-	c.Assert(len(d), Equals, 1)
-}
-
-func (ts *HTestSuite) TestAtomicWriteFilePermissions(c *C) {
-	tmpdir := c.MkDir()
-
-	p := filepath.Join(tmpdir, "foo")
-	err := AtomicWriteFile(p, []byte(""), 0600, 0)
-	c.Assert(err, IsNil)
-
-	st, err := os.Stat(p)
-	c.Assert(err, IsNil)
-	c.Assert(st.Mode()&os.ModePerm, Equals, os.FileMode(0600))
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileOverwrite(c *C) {
-	tmpdir := c.MkDir()
-	p := filepath.Join(tmpdir, "foo")
-	c.Assert(ioutil.WriteFile(p, []byte("hello"), 0644), IsNil)
-	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, 0), IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "hi")
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileSymlinkNoFollow(c *C) {
-	tmpdir := c.MkDir()
-	rodir := filepath.Join(tmpdir, "ro")
-	p := filepath.Join(rodir, "foo")
-	s := filepath.Join(tmpdir, "foo")
-	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
-	c.Assert(os.Symlink(s, p), IsNil)
-	c.Assert(os.Chmod(rodir, 0500), IsNil)
-	defer os.Chmod(rodir, 0700)
-
-	err := AtomicWriteFile(p, []byte("hi"), 0600, 0)
-	c.Assert(err, NotNil)
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileAbsoluteSymlinks(c *C) {
-	tmpdir := c.MkDir()
-	rodir := filepath.Join(tmpdir, "ro")
-	p := filepath.Join(rodir, "foo")
-	s := filepath.Join(tmpdir, "foo")
-	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
-	c.Assert(os.Symlink(s, p), IsNil)
-	c.Assert(os.Chmod(rodir, 0500), IsNil)
-	defer os.Chmod(rodir, 0700)
-
-	err := AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow)
-	c.Assert(err, IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "hi")
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileOverwriteAbsoluteSymlink(c *C) {
-	tmpdir := c.MkDir()
-	rodir := filepath.Join(tmpdir, "ro")
-	p := filepath.Join(rodir, "foo")
-	s := filepath.Join(tmpdir, "foo")
-	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
-	c.Assert(os.Symlink(s, p), IsNil)
-	c.Assert(os.Chmod(rodir, 0500), IsNil)
-	defer os.Chmod(rodir, 0700)
-
-	c.Assert(ioutil.WriteFile(s, []byte("hello"), 0644), IsNil)
-	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow), IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "hi")
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileRelativeSymlinks(c *C) {
-	tmpdir := c.MkDir()
-	rodir := filepath.Join(tmpdir, "ro")
-	p := filepath.Join(rodir, "foo")
-	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
-	c.Assert(os.Symlink("../foo", p), IsNil)
-	c.Assert(os.Chmod(rodir, 0500), IsNil)
-	defer os.Chmod(rodir, 0700)
-
-	err := AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow)
-	c.Assert(err, IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "hi")
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileOverwriteRelativeSymlink(c *C) {
-	tmpdir := c.MkDir()
-	rodir := filepath.Join(tmpdir, "ro")
-	p := filepath.Join(rodir, "foo")
-	s := filepath.Join(tmpdir, "foo")
-	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
-	c.Assert(os.Symlink("../foo", p), IsNil)
-	c.Assert(os.Chmod(rodir, 0500), IsNil)
-	defer os.Chmod(rodir, 0700)
-
-	c.Assert(ioutil.WriteFile(s, []byte("hello"), 0644), IsNil)
-	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow), IsNil)
-
-	content, err := ioutil.ReadFile(p)
-	c.Assert(err, IsNil)
-	c.Assert(string(content), Equals, "hi")
-}
-
-func (ts *HTestSuite) TestAtomicWriteFileNoOverwriteTmpExisting(c *C) {
-	tmpdir := c.MkDir()
-	// ensure we always get the same result
-	rand.Seed(1)
-	expectedRandomness := MakeRandomString(12)
-	// ensure we always get the same result
-	rand.Seed(1)
-
-	p := filepath.Join(tmpdir, "foo")
-	err := ioutil.WriteFile(p+"."+expectedRandomness, []byte(""), 0644)
-	c.Assert(err, IsNil)
-
-	err = AtomicWriteFile(p, []byte(""), 0600, 0)
-	c.Assert(err, ErrorMatches, "open .*: file exists")
-}
-
-func (ts *HTestSuite) TestCurrentHomeDirHOMEenv(c *C) {
-	tmpdir := c.MkDir()
-
-	oldHome := os.Getenv("HOME")
-	defer os.Setenv("HOME", oldHome)
-
-	os.Setenv("HOME", tmpdir)
-	home, err := CurrentHomeDir()
-	c.Assert(err, IsNil)
-	c.Assert(home, Equals, tmpdir)
-}
-
-func (ts *HTestSuite) TestCurrentHomeDirNoHomeEnv(c *C) {
-	oldHome := os.Getenv("HOME")
-	if oldHome == "/sbuild-nonexistent" {
-		c.Skip("running in schroot this test won't work")
-	}
-	defer os.Setenv("HOME", oldHome)
-
-	os.Setenv("HOME", "")
-	home, err := CurrentHomeDir()
-	c.Assert(err, IsNil)
-	c.Assert(home, Equals, oldHome)
-}
-
 func skipOnMissingDevKmsg(c *C) {
 	_, err := os.Stat("/dev/kmsg")
 	if err != nil {
