+++ osutil/io_test.go
@@ -1,0 +1,176 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2014-2015 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package osutil
+
+import (
+	"io/ioutil"
+	"math/rand"
+	"os"
+	"path/filepath"
+
+	"github.com/ubuntu-core/snappy/helpers"
+
+	. "gopkg.in/check.v1"
+)
+
+type AtomicWriteTestSuite struct{}
+
+var _ = Suite(&AtomicWriteTestSuite{})
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFile(c *C) {
+	tmpdir := c.MkDir()
+
+	p := filepath.Join(tmpdir, "foo")
+	err := AtomicWriteFile(p, []byte("canary"), 0644, 0)
+	c.Assert(err, IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "canary")
+
+	// no files left behind!
+	d, err := ioutil.ReadDir(tmpdir)
+	c.Assert(err, IsNil)
+	c.Assert(len(d), Equals, 1)
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFilePermissions(c *C) {
+	tmpdir := c.MkDir()
+
+	p := filepath.Join(tmpdir, "foo")
+	err := AtomicWriteFile(p, []byte(""), 0600, 0)
+	c.Assert(err, IsNil)
+
+	st, err := os.Stat(p)
+	c.Assert(err, IsNil)
+	c.Assert(st.Mode()&os.ModePerm, Equals, os.FileMode(0600))
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileOverwrite(c *C) {
+	tmpdir := c.MkDir()
+	p := filepath.Join(tmpdir, "foo")
+	c.Assert(ioutil.WriteFile(p, []byte("hello"), 0644), IsNil)
+	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, 0), IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "hi")
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileSymlinkNoFollow(c *C) {
+	tmpdir := c.MkDir()
+	rodir := filepath.Join(tmpdir, "ro")
+	p := filepath.Join(rodir, "foo")
+	s := filepath.Join(tmpdir, "foo")
+	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
+	c.Assert(os.Symlink(s, p), IsNil)
+	c.Assert(os.Chmod(rodir, 0500), IsNil)
+	defer os.Chmod(rodir, 0700)
+
+	err := AtomicWriteFile(p, []byte("hi"), 0600, 0)
+	c.Assert(err, NotNil)
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileAbsoluteSymlinks(c *C) {
+	tmpdir := c.MkDir()
+	rodir := filepath.Join(tmpdir, "ro")
+	p := filepath.Join(rodir, "foo")
+	s := filepath.Join(tmpdir, "foo")
+	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
+	c.Assert(os.Symlink(s, p), IsNil)
+	c.Assert(os.Chmod(rodir, 0500), IsNil)
+	defer os.Chmod(rodir, 0700)
+
+	err := AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow)
+	c.Assert(err, IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "hi")
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileOverwriteAbsoluteSymlink(c *C) {
+	tmpdir := c.MkDir()
+	rodir := filepath.Join(tmpdir, "ro")
+	p := filepath.Join(rodir, "foo")
+	s := filepath.Join(tmpdir, "foo")
+	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
+	c.Assert(os.Symlink(s, p), IsNil)
+	c.Assert(os.Chmod(rodir, 0500), IsNil)
+	defer os.Chmod(rodir, 0700)
+
+	c.Assert(ioutil.WriteFile(s, []byte("hello"), 0644), IsNil)
+	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow), IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "hi")
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileRelativeSymlinks(c *C) {
+	tmpdir := c.MkDir()
+	rodir := filepath.Join(tmpdir, "ro")
+	p := filepath.Join(rodir, "foo")
+	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
+	c.Assert(os.Symlink("../foo", p), IsNil)
+	c.Assert(os.Chmod(rodir, 0500), IsNil)
+	defer os.Chmod(rodir, 0700)
+
+	err := AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow)
+	c.Assert(err, IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "hi")
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileOverwriteRelativeSymlink(c *C) {
+	tmpdir := c.MkDir()
+	rodir := filepath.Join(tmpdir, "ro")
+	p := filepath.Join(rodir, "foo")
+	s := filepath.Join(tmpdir, "foo")
+	c.Assert(os.MkdirAll(rodir, 0755), IsNil)
+	c.Assert(os.Symlink("../foo", p), IsNil)
+	c.Assert(os.Chmod(rodir, 0500), IsNil)
+	defer os.Chmod(rodir, 0700)
+
+	c.Assert(ioutil.WriteFile(s, []byte("hello"), 0644), IsNil)
+	c.Assert(AtomicWriteFile(p, []byte("hi"), 0600, AtomicWriteFollow), IsNil)
+
+	content, err := ioutil.ReadFile(p)
+	c.Assert(err, IsNil)
+	c.Assert(string(content), Equals, "hi")
+}
+
+func (ts *AtomicWriteTestSuite) TestAtomicWriteFileNoOverwriteTmpExisting(c *C) {
+	tmpdir := c.MkDir()
+	// ensure we always get the same result
+	rand.Seed(1)
+	expectedRandomness := helpers.MakeRandomString(12)
+	// ensure we always get the same result
+	rand.Seed(1)
+
+	p := filepath.Join(tmpdir, "foo")
+	err := ioutil.WriteFile(p+"."+expectedRandomness, []byte(""), 0644)
+	c.Assert(err, IsNil)
+
+	err = AtomicWriteFile(p, []byte(""), 0600, 0)
+	c.Assert(err, ErrorMatches, "open .*: file exists")
+}
