+++ snappy/firstboot.go
@@ -26,7 +26,6 @@ import (
 	"os/exec"
 	"path/filepath"
 
-	"github.com/ubuntu-core/snappy/helpers"
 	"github.com/ubuntu-core/snappy/logger"
 	"github.com/ubuntu-core/snappy/osutil"
 	"github.com/ubuntu-core/snappy/progress"
@@ -190,7 +189,7 @@ func stampFirstBoot() error {
 		}
 	}
 
-	return helpers.AtomicWriteFile(stampFile, []byte{}, 0644, 0)
+	return osutil.AtomicWriteFile(stampFile, []byte{}, 0644, 0)
 }
 
 var globs = []string{"/sys/class/net/eth*", "/sys/class/net/en*"}
@@ -217,7 +216,7 @@ func enableFirstEther() error {
 	ethfile := filepath.Join(ethdir, eth)
 	data := fmt.Sprintf("allow-hotplug %[1]s\niface %[1]s inet dhcp\n", eth)
 
-	if err := helpers.AtomicWriteFile(ethfile, []byte(data), 0644, 0); err != nil {
+	if err := osutil.AtomicWriteFile(ethfile, []byte(data), 0644, 0); err != nil {
 		return err
 	}
